database/pyproject.toml
[tool.poetry]
name = "static-website-database-compiler"
version = "0.1.0"
description = "Compiles a YAML database to JSON for use with generated static websites"
authors = ["Ewen Le Bihan <ewen.lebihan7@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pastel = "^0.2.0"
toml = "^0.10.0"
docopt = "^0.6.2"
yaspin = "^0.16.0"
progressbar = "^2.5"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
